feat(test): add dummy slack client in new tests
