+++ apps/project/tests/mutation_test.py
@@ -7,12 +7,11 @@ from unittest.mock import call, patch
 from django.conf import settings
 from django.core.files.temp import NamedTemporaryFile
 from PIL import Image
-from slack_sdk.models.blocks.blocks import Block
-from slack_sdk.web.slack_response import SlackResponse
 from ulid import ULID
 
 from apps.common.models import IconEnum
 from apps.contributor.factories import ContributorTeamFactory
+from apps.project.dummy_slack_client import MapswipeSlackMock
 from apps.project.factories import OrganizationFactory, ProjectFactory
 from apps.project.models import (
     Organization,
@@ -32,7 +31,6 @@ from main.tests import TestCase
 from project_types.street import project as street_project
 from project_types.tile_map_service.compare import project as compare_project
 from utils.geo.raster_tile_server.config import RasterTileServerNameEnum
-from utils.slack import MapswipeSlack
 
 BASE_DIR = Path(__file__).resolve().parent
 
@@ -111,49 +109,6 @@ def update_project_query(
         },
         **kwargs,
     )
-
-
-class MapswipeSlackMock(MapswipeSlack):
-    slack_thread_ts_counter = 0
-
-    class DummySlackClient:
-        def chat_postMessage(self, **kwargs):
-            # Return a fake Slack response
-            return {"ok": True, "ts": "FAKE_TS"}
-
-        def chat_update(self, **kwargs):
-            return {"ok": True, "ts": "FAKE TS"}
-
-    @typing.override
-    def __init__(self):
-        self.client = self.DummySlackClient()
-
-    @typing.override
-    def send_slack_message(
-        self,
-        text: str | None = None,
-        blocks: str | typing.Sequence[dict | Block] | None = None,
-        thread_ts: str | None = None,
-        reply_broadcast: bool = True,
-    ) -> SlackResponse:
-        self.slack_thread_ts_counter += 1
-        return typing.cast(
-            "SlackResponse",
-            {"ts": f"{self.slack_thread_ts_counter}.7890"},
-        )
-
-    @typing.override
-    def update_slack_message(
-        self,
-        ts: str,
-        text: str | None = None,
-        blocks: typing.Sequence[dict | Block] | None = None,
-    ) -> SlackResponse:
-        self.slack_thread_ts_counter += 1
-        return typing.cast(
-            "SlackResponse",
-            {"ts": f"{self.slack_thread_ts_counter}.7890"},
-        )
 
 
 class Mutation:
